# api.yaml
openapi: 3.0.1
info:
  title: Match Reservation API
  version: 0.1.0
  description: API for Match Reservation Website
servers:
- url: http://localhost:3000/
  description: local development server
paths:
  /api/users:
    post:
      summary: signup
      operationId: signup
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/SubmittedUser'
      responses:
        '200':
          description: ok
          content:
            'application/json':
              schema:
                type: object
                description: successful signup
                properties:
                  authToken:
                    type: string
                  userType:
                    type: string
                    enum:
                    - fan
                    - manager
                    - admin
                    description: used by the forntend to know the type of the user, the backend still needs to check the payload of the authToken payload
                  msg:
                    type: string
                required:
                - authToken
                - userType
                - msg

        '401':
          description: not authorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: invalid data
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: get all users
      operationId: getUsers
      parameters:
        - name: x-auth-token
          in: header
          schema:
            type: string
          required: true
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: ok
          content:
            'application/json':
              schema:
                type: object
                properties:
                  has_next:
                    type: boolean
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/ListedUser'
                required:
                - has_next
                - users
        '406':
          description: not acceptable, if page < 1
  /api/users/me:
    get:
      summary: get my info
      operationId: getMyInfo
      parameters:
        - name: x-auth-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ListedUser'
        '401':
          description: not authorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: edit user
      operationId: editUser
      parameters:
        - name: x-auth-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/EditedUser'
      responses:
        '200':
          description: ok
          content:
            'application/json':
              schema:
                type: object
                properties:
                  logout:
                    type: boolean
                  user:
                    type: object
                    items:
                      $ref: '#/components/schemas/ListedUser'
                  msg:
                    type: string
                    description: in case the user asked to be a manger, send a message to display to assure they will be reviewed
                required:
                - logout
                - user
        '401':
          description: not authorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{username}:
    delete:
      summary: delete a user
      operationId: deleteUser
      parameters:
        - name: username
          required: true
          in: path
          schema:
            type: string
        - name: x-auth-token
          in: header
          schema:
            type: string
          description: must be an admin, or the account owner
          required: true
      responses:
        '200':
          description: ok
        '401':
          description: not authorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: username not found
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/matches:
    get:
      summary: get matches
      operationId: getMatches
      parameters:
        - name: page
          in: query
          required: false
          schema:
            default: 1
            type: integer
      responses:
        '200':
          description: ok
          content:
            'application/json':
              schema:
                type: object
                properties:
                  has_next:
                    type: boolean
                  matches:
                    type: array
                    items:
                      $ref: '#/components/schemas/ListedMatch'
                required:
                - has_next
                - matches
        '406':
          description: not acceptable, if page < 1
    post:
      summary: submit a match
      operationId: submitMatch
      parameters:
        - name: x-auth-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/SubmittedMatch'
      responses:
        '200':
          description: ok
        '401':
          description: not authorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: invalid data
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/matches/{match_id}:
    get:
      summary: get one match
      operationId: getMatch
      parameters:
        - name: match_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ListedMatch'
        '404':
          description: match not found
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: edit a match
      operationId: editMatch
      parameters:
        - name: x-auth-token
          in: header
          required: true
          schema:
            type: string
        - name: match_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/SubmittedMatch'
      responses:
        '200':
          description: ok
        '401':
          description: not authorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: invalid data
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: delete a match
      operationId: deleteMatch
      parameters:
        - name: x-auth-token
          in: header
          required: true
          schema:
            type: string
        - name: match_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
        '404':
          description: match not found
        '401':
          description: not authorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/matches/{match_id}/seats:
    get:
      summary: get seats of given match
      operationId: getSeats
      parameters:
        - name: match_id
          schema:
            type: string
          required: true
          in: path
        - name: x-auth-token
          schema:
            type: string
          required: false
          in: header
      responses:
        '200':
          description: ok
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Seats'
        '404':
          description: match not found
  /api/matches/{match_id}/seats/reserve/{seat_id}:
    post:
      summary: reserve seat
      operationId: reserveSeat
      parameters:
        - name: match_id
          schema:
            type: string
          required: true
          in: path
        - name: seat_id
          schema:
            type: string
          required: true
          in: path
        - name: x-auth-token
          schema:
            type: string
          required: true
          in: header
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/CreditCard'
      responses:
        '200':
          description: ok
        '404':
          description: match/seat not found
        '400':
          description: invalid credit card details
        '401':
          description: not authorized, or reserved
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    SubmittedUser:
      type: object
      additionalProperties: false
      description: User sends this object when they first signup
      properties:
        username:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          example: '2020-12-09'
          type: string
        gender:
          type: string
          enum:
          - male
          - female
        city:
          type: string
        address:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
          - fan
          - manager
      required:
      - username
      - password
      - firstName
      - lastName
      - birthDate
      - gender
      - city
      - address
      - email
      - role
    EditedUser:
      type: object
      additionalProperties: false
      description: User sends this object when they edit their profile
      properties:
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          example: '2020-12-09'
          type: string
        gender:
          type: string
          enum:
          - male
          - female
        city:
          type: string
        address:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
          - fan
          - manager
    ListedUser:
      type: object
      additionalProperties: false
      description: This is the object sent to frontend when it asks to list user
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
        gender:
          type: string
          enum:
          - male
          - female
        city:
          type: string
        address:
          type: string
        email:
          type: string
      required:
      - username
      - firstName
      - lastName
      - birthDate
      - gender
      - city
      - address
      - email
    ErrorResponse:
      type: object
      additionalProperties: true
      description: when an error happens, this holds the error description
      properties:
        err:
          type: string
      required:
      - err
    SubmittedMatch:
      type: object
      additionalProperties: false
      properties:
        homeTeam:
          type: string
        awayTeam:
          type: string
        venue:
          type: string
        dateTime:
          example: '2020-12-09T23:12:00+00:00'
          type: string
        mainReferee:
          type: string
        firstLinesman:
          type: string
        secondLinesman:
          type: string
      required:
      - homeTeam
      - awayTeam
      - venue
      - dateTime
      - mainReferee
      - firstLinesman
      - secondLinesman
    ListedMatch:
      type: object
      additionalProperties: false
      properties:
        homeTeam:
          type: string
        awayTeam:
          type: string
        venue:
          type: string
        dateTime:
          type: string
        mainReferee:
          type: string
        firstLinesman:
          type: string
        secondLinesman:
          type: string
        ticketPrice:
          type: integer
        uuid:
          type: string
      required:
      - homeTeam
      - awayTeam
      - venue
      - dateTime
      - mainReferee
      - firstLinesman
      - secondLinesman
      - ticketPrice
      - uuid
    Seats:
      type: object
      additionalProperties: false
...
    CreditCard:
      type: object
      additionalProperties: false
      properties:
        creditCardNumber:
          type: string
        pin:
          type: string
      required:
      - creditCardNumber
      - pin